pyproject.toml
[project]
name = "ai-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
